move output formatting to output utils
